Job submission now functional with callbacks. Notifications should be working too.
