+++ src/index.html
@@ -52,6 +52,7 @@
     <script src="js/WelcomeCtrl.js"></script>
     <script src="js/DebugLogger.js"></script>
     <script src="js/ConnectionService.js"></script>
+    <script src="js/filePathService.js"></script>
     <script src="js/clusterLandingCtrl.js"></script>
     <script src="js/jobSubmissionCtrl.js"></script>
     <script src="js/jobHistoryCtrl.js"></script>
